refactor(structure): move to feature-oriented layout
Group 4 of rework-state-architecture (isolated, behavior-preserving).
Relocate files into the layout from design D7:
- src/app/: store infra — store.ts, hooks.ts, listenerMiddleware.ts,
  githubApi.ts (+ their tests), moved from src/store/ (removed).
- src/features/searchHistory/: searchHistorySlice.ts (+ test) and the
  Form, ListRequests, AccordionItem, Repository component folders.
- src/components/: only shared dumb UI (Header, Loader) remains.

Imports change by name via the src path alias ("*": ["./src/*"]):
store/X → app/X, store/searchHistorySlice and the four components →
features/searchHistory/…. No feature-to-feature import edge is
introduced — githubApi is a shared service in app/, not a peer feature.

Validated: typecheck, lint, 36 tests, coverage 100% — unchanged from
group 3 (no behavior change). Gate 3 code-review: clean. No test logic
changed, so Gate 4 N/A.
