+++ src/features/searchHistory/AccordionItem/styles.module.scss
