+++ src/features/searchHistory/Repository/styles.module.scss
